+++ NEWS.md
@@ -4,6 +4,7 @@ New arguments to `expect_snapshot_print()`:
 
 * `fn_current`: A function to process the current snapshot string before comparison.
 * `fn_target`: A function to process the target snapshot string before comparison.
+* `review`: TRUE (default) to write the diff plot to file. FALSE to return a failure but no file.
 
 # 0.0.8